Add Registration and Authentication Components
For our web app to be usable , we need users to signup and login.
Signup creates an account , login provides access to the dashboard.
To manage this , we designate some route as protected by
housing such routes inside an array

Request to any route checks for a valid user token before the user can
access vital services